refactor(metadata): centralize subjective dimension catalog and remove scoring cycle
